Change FE to use `long` with 64-bit support

+++ src/main/java/net/neoforged/neoforge/energy/EmptyEnergyStorage.java
@@ -21,22 +21,22 @@ public class EmptyEnergyStorage implements IEnergyStorage {
     protected EmptyEnergyStorage() {}
 
     @Override
-    public int receiveEnergy(int maxReceive, boolean simulate) {
+    public long receiveEnergy(long maxReceive, boolean simulate) {
         return 0;
     }
 
     @Override
-    public int extractEnergy(int maxExtract, boolean simulate) {
+    public long extractEnergy(long maxExtract, boolean simulate) {
         return 0;
     }
 
     @Override
-    public int getEnergyStored() {
+    public long getEnergyStored() {
         return 0;
     }
 
     @Override
-    public int getMaxEnergyStored() {
+    public long getMaxEnergyStored() {
         return 0;
     }
